+++ src/submissions/frontend/custom_widgets/misc.py
@@ -2,20 +2,23 @@
 Contains miscellaneous widgets for frontend functions
 '''
 from datetime import date
-import typing
+import difflib
+from typing import Tuple
 from PyQt6.QtWidgets import (
     QLabel, QVBoxLayout,
     QLineEdit, QComboBox, QDialog, 
     QDialogButtonBox, QDateEdit, QSizePolicy, QWidget,
     QGridLayout, QPushButton, QSpinBox, QDoubleSpinBox,
-    QHBoxLayout,
+    QHBoxLayout, QMainWindow
 )
 from PyQt6.QtCore import Qt, QDate, QSize
 # from submissions.backend.db.functions import lookup_kittype_by_use
 # from submissions.backend.db import lookup_regent_by_type_name_and_kit_name
 from tools import check_not_nan
 from ..all_window_functions import extract_form_info
-from backend.db import get_all_reagenttype_names, lookup_all_sample_types, create_kit_from_yaml, lookup_regent_by_type_name, lookup_kittype_by_use#, lookup_regent_by_type_name_and_kit_name
+from backend.db import get_all_reagenttype_names, lookup_all_sample_types, create_kit_from_yaml, \
+    lookup_regent_by_type_name, lookup_kittype_by_use, lookup_all_orgs
+    #, lookup_regent_by_type_name_and_kit_name
 from backend.excel.parser import SheetParser
 from jinja2 import Environment, FileSystemLoader
 import sys
@@ -297,55 +300,103 @@ class ControlsDatePicker(QWidget):
         return QSize(80,20)  
 
 
+# class ImportReagent(QComboBox):
+
+#     def __init__(self, ctx:dict, item:str, prsr:SheetParser|None=None):
+#         super().__init__()
+#         self.setEditable(True)
+#         # Ensure that all reagenttypes have a name that matches the items in the excel parser
+#         query_var = item.replace("lot_", "")
+#         if prsr != None:
+#             logger.debug(f"Import Reagent is looking at: {prsr.sub[item]} for {item}")
+#         else:
+#             logger.debug(f"Import Reagent is going to retrieve all reagents for {item}")
+#         logger.debug(f"Query for: {query_var}")
+#         if prsr != None:
+#             if isinstance(prsr.sub[item], np.float64):
+#                 logger.debug(f"{prsr.sub[item]['lot']} is a numpy float!")
+#                 try:
+#                     prsr.sub[item] = int(prsr.sub[item]['lot'])
+#                 except ValueError:
+#                     pass
+#         # query for reagents using type name from sheet and kit from sheet
+#         logger.debug(f"Attempting lookup of reagents by type: {query_var}")
+#         # below was lookup_reagent_by_type_name_and_kit_name, but I couldn't get it to work.
+#         relevant_reagents = [item.__str__() for item in lookup_regent_by_type_name(ctx=ctx, type_name=query_var)]#, kit_name=prsr.sub['extraction_kit'])]
+#         # relevant_reagents = [item.__str__() for item in lookup_regent_by_type_name_and_kit_name(ctx=ctx, type_name=query_var, kit_name=prsr.sub['extraction_kit'])]
+#         output_reg = []
+#         for reagent in relevant_reagents:
+#             # extract strings from any sets.
+#             if isinstance(reagent, set):
+#                 for thing in reagent:
+#                     output_reg.append(thing)
+#             elif isinstance(reagent, str):
+#                 output_reg.append(reagent)
+#         relevant_reagents = output_reg
+#         # if reagent in sheet is not found insert it into the front of relevant reagents so it shows 
+#         if prsr != None:
+#             logger.debug(f"Relevant reagents for {prsr.sub[item]}: {relevant_reagents}")
+#             if str(prsr.sub[item]['lot']) not in relevant_reagents:
+#                 if check_not_nan(prsr.sub[item]['lot']):
+#                     relevant_reagents.insert(0, str(prsr.sub[item]['lot']))
+#             else:
+#                 if len(relevant_reagents) > 1:
+#                     logger.debug(f"Found {prsr.sub[item]['lot']} in relevant reagents: {relevant_reagents}. Moving to front of list.")
+#                     idx = relevant_reagents.index(str(prsr.sub[item]['lot']))
+#                     logger.debug(f"The index we got for {prsr.sub[item]['lot']} in {relevant_reagents} was {idx}")
+#                     moved_reag = relevant_reagents.pop(idx)
+#                     relevant_reagents.insert(0, moved_reag)
+#                 else:
+#                     logger.debug(f"Found {prsr.sub[item]['lot']} in relevant reagents: {relevant_reagents}. But no need to move due to short list.")
+#         logger.debug(f"New relevant reagents: {relevant_reagents}")
+#         self.setObjectName(f"lot_{item}")
+#         self.addItems(relevant_reagents)
+        
+
 class ImportReagent(QComboBox):
 
-    def __init__(self, ctx:dict, item:str, prsr:SheetParser|None=None):
+    def __init__(self, ctx:dict, reagent:dict):
         super().__init__()
         self.setEditable(True)
         # Ensure that all reagenttypes have a name that matches the items in the excel parser
-        query_var = item.replace("lot_", "")
-        if prsr != None:
-            logger.debug(f"Import Reagent is looking at: {prsr.sub[item]} for {item}")
-        else:
-            logger.debug(f"Import Reagent is going to retrieve all reagents for {item}")
-        logger.debug(f"Query for: {query_var}")
-        if prsr != None:
-            if isinstance(prsr.sub[item], np.float64):
-                logger.debug(f"{prsr.sub[item]['lot']} is a numpy float!")
-                try:
-                    prsr.sub[item] = int(prsr.sub[item]['lot'])
-                except ValueError:
-                    pass
+        query_var = reagent['type']
+        logger.debug(f"Import Reagent is looking at: {reagent['lot']} for {reagent['type']}")
+        
+        if isinstance(reagent['lot'], np.float64):
+            logger.debug(f"{reagent['lot']} is a numpy float!")
+            try:
+                reagent['lot'] = int(reagent['lot'])
+            except ValueError:
+                pass
         # query for reagents using type name from sheet and kit from sheet
         logger.debug(f"Attempting lookup of reagents by type: {query_var}")
         # below was lookup_reagent_by_type_name_and_kit_name, but I couldn't get it to work.
         relevant_reagents = [item.__str__() for item in lookup_regent_by_type_name(ctx=ctx, type_name=query_var)]#, kit_name=prsr.sub['extraction_kit'])]
         # relevant_reagents = [item.__str__() for item in lookup_regent_by_type_name_and_kit_name(ctx=ctx, type_name=query_var, kit_name=prsr.sub['extraction_kit'])]
         output_reg = []
-        for reagent in relevant_reagents:
+        for rel_reagent in relevant_reagents:
             # extract strings from any sets.
-            if isinstance(reagent, set):
-                for thing in reagent:
+            if isinstance(rel_reagent, set):
+                for thing in rel_reagent:
                     output_reg.append(thing)
-            elif isinstance(reagent, str):
-                output_reg.append(reagent)
+            elif isinstance(rel_reagent, str):
+                output_reg.append(rel_reagent)
         relevant_reagents = output_reg
         # if reagent in sheet is not found insert it into the front of relevant reagents so it shows 
-        if prsr != None:
-            logger.debug(f"Relevant reagents for {prsr.sub[item]}: {relevant_reagents}")
-            if str(prsr.sub[item]['lot']) not in relevant_reagents:
-                if check_not_nan(prsr.sub[item]['lot']):
-                    relevant_reagents.insert(0, str(prsr.sub[item]['lot']))
+        # if prsr != None:
+        logger.debug(f"Relevant reagents for {reagent['lot']}: {relevant_reagents}")
+        if str(reagent['lot']) not in relevant_reagents:
+            if check_not_nan(reagent['lot']):
+                relevant_reagents.insert(0, str(reagent['lot']))
+        else:
+            if len(relevant_reagents) > 1:
+                logger.debug(f"Found {reagent['lot']} in relevant reagents: {relevant_reagents}. Moving to front of list.")
+                idx = relevant_reagents.index(str(reagent['lot']))
+                logger.debug(f"The index we got for {reagent['lot']} in {relevant_reagents} was {idx}")
+                moved_reag = relevant_reagents.pop(idx)
+                relevant_reagents.insert(0, moved_reag)
             else:
-                if len(relevant_reagents) > 1:
-                    logger.debug(f"Found {prsr.sub[item]['lot']} in relevant reagents: {relevant_reagents}. Moving to front of list.")
-                    idx = relevant_reagents.index(str(prsr.sub[item]['lot']))
-                    logger.debug(f"The index we got for {prsr.sub[item]['lot']} in {relevant_reagents} was {idx}")
-                    moved_reag = relevant_reagents.pop(idx)
-                    relevant_reagents.insert(0, moved_reag)
-                else:
-                    logger.debug(f"Found {prsr.sub[item]['lot']} in relevant reagents: {relevant_reagents}. But no need to move due to short list.")
+                logger.debug(f"Found {reagent['lot']} in relevant reagents: {relevant_reagents}. But no need to move due to short list.")
         logger.debug(f"New relevant reagents: {relevant_reagents}")
-        self.setObjectName(f"lot_{item}")
+        self.setObjectName(f"lot_{reagent['type']}")
         self.addItems(relevant_reagents)
-